Change license to Apache 2.0
We have decided to change license to Apache 2.0, which is a permissive open source license.
See https://opensource.org/faq#permissive

We do this change to make camacq more compatible with different contexts that may require a
permissive license. We have chosen Apache 2.0 since it explicitly says that all contributions
should be included under the same license as the current work. This makes a separate CLA
less needed.